refactor: LobbyControllerAdvice 추가
